Add files via upload
This is an Arduino program that takes Potentiometer and Button data from the master arm and sends it over serial USB for the Up-board, computer, or Raspberry Pi to use. (at the time I'm writing this the use is to send the data over WiFi to the rover which will move the slave arm accordingly. At this point the program has not been tested because there isn't a arm to test it on.