Add files via upload
(The other pull request was to create a folder to put these files in.)

This code reads in a .csv file with the average terrain values (created in MATLAB) and attaches GPS coordinates to it. Currently the averages are of 100 pixels. Attempts to do for each pixel are currently running into stack overflow problems.