Arm PIDs - not so rough and now in .ino form
Arm PIDs - one each for base, shoulder, elbow, wrist pitch, and wrist roll
Each one takes in a target position between 0 and 4095, and moves the motor of the joint to reach that value


Things still to be done:
Add pin numbers for the motors and potentiometers
Function in loop that receives the message char array and gives to adjust arm method
Tune the PIDs - currently they are only Ps w/ value of 1
What value is the potentiometer outputting - currently it assumes an angle, i.e. 0 to 270
What value is written to the servo - currently it assumes an angle, i.e. 0 to 270

Changes from previous versions:
Changed to .ino form
Fully fleshed out the PID for each joint
Converts from 2 chars to degrees