Add files via upload
Same changes as the previous pull request (the .ino one), with the addition of fixing a position vs angle problem. Also should now properly convert the angle read in from the potentiometer.
(Angle is 0 to 180 - so not really angle, but 0 is full back and 180 is full forward. Position is the rough number, so out of 4096 or 1024).